Use root d files for TS resolution + move CDS types into folder
