+++ src/types/index.d.ts
@@ -1,2 +1,1 @@
-import "./cds-hooks/index.d.ts"
 import "./smart/index.d.ts"
